+++ src/types/post.ts
@@ -1,0 +1,3 @@
+export type PostsProps = {
+  topics: string[],
+};
